Create compression.c
This program uses an extremely basic version of algorithmic coding to compress the image. Note that the Y channel is compressed in a higher quality compared to the other two channels ( Luma preference ). This program takes the data from files named Y.pgm, Cb.pgm and Cr.pgm, which contain the data from the processed image, compresses each of them separately, recombines them and stores the compressed image in an output file names comp.ppm